pyproject.toml
[project]
name = "switchboard"
version = "0.1.0"
description = "managed loadbalancing for azure openai"
readme = "README.md"
authors = [
    { name = "abizer", email = "abizer@abizer.me" }
]
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.8.0",
    "openai>=1.62.0",
    "opentelemetry-api>=1.30.0",
    "tenacity>=9.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
    "ipython>=8.0.0,!=9.0.0",
    "pytest>=8.3.4",
    "pytest-anyio>=0.0.0",
    "pytest-asyncio>=0.25.3",
    "ruff>=0.9.6",
]
